pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medderm"
version = "0.1.0"
description = "Dermatology model & XAI toolkit with saliency, FastAPI, and Gradio UI"
authors = [
    { name = "Leonardo Sgroi", email = "lsgroi@studenti.uninsubria.it"},
]
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    # Core runtime
    "numpy==1.26.4",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "scipy==1.15.2",
    "scikit-image==0.25.2",
    "matplotlib==3.10.1",

    # DL / Vision
    "torch==2.6.0",
    "torchvision==0.21.0",
    "timm==1.0.15",
    "transformers==4.51.3",
    "tokenizers==0.21.1",
    "safetensors==0.5.3",
    "huggingface-hub==0.30.2",

    # XAI
    "captum==0.8.0",

    # Web/API & UI
    "fastapi==0.115.12",
    "uvicorn==0.34.1",
    "gradio==5.25.2",
    "gradio_client==1.8.0",
    "python-multipart==0.0.20",
    "jinja2==3.1.6",
    "aiofiles==24.1.0",

    # LLM / LangChain stack
    "langchain-core==0.3.52",
    "langchain-openai==0.3.13",
    "langgraph==0.3.30",
    "langgraph-checkpoint==2.0.24",
    "langgraph-prebuilt==0.1.8",
    "langgraph-sdk==0.1.61",
    "langsmith==0.3.31",
    "openai==1.74.0",
    "tiktoken==0.9.0",

    # Utils
    "python-dotenv==1.1.0",
    "pydantic==2.11.3",
    "orjson==3.10.16",
    "requests==2.32.3",
    "httpx==0.28.1",
    "anyio==4.9.0",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "rich==14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff==0.11.5",
    "black",
    "isort",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["medDerm*"]
exclude = ["tests*", "demo*", "examples*", "med_env*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]